fix: fix height and wieght types

+++ Tenisu.Domain/Data.cs
@@ -4,8 +4,8 @@
     {
         public int Rank { get; set; }
         public int Points { get; set; }
-        public int Weight { get; set; }
-        public int Height { get; set; }
+        public double Weight { get; set; }
+        public double Height { get; set; }
         public int Age { get; set; }
         public List<double> Last { get; set; }
     }
